+++ lib/colors.ts
@@ -1,4 +1,5 @@
 import type { Color, Oklch, Rgb, Lch } from 'culori/fn'
+
 import {
   formatRgb as formatRgbFast,
   parse as originParse,
