+++ stores/formats.ts
@@ -1,4 +1,5 @@
 import type { OutputFormats } from './settings.js'
+
 import {
   formatHex8,
   formatHex,
