+++ stores/visible.ts
@@ -1,4 +1,5 @@
 import type { Color } from 'culori/fn'
+
 import { computed } from 'nanostores'
 
 import {
